Implement guess check and guess digit check methods in breaker class
The guess check method checks whether the player guess is equal to the
code created by the code maker or not. The guess digit check method
individually checks every digit in the player guess if the digit is
found in the same position, or a different position or the digit is
not found in the code created by the code maker.