Implement player choice in code creation in maker class
The choose type method checks the player type (e.g. Human or Computer),
and provides the appropriate response per player type. The Computer
player gets to randomize a 4-digit code, while the Human player gets
to choose whether to randomize the 4-digit code or to manually input
the 4-digit code.