Implement game loop methods in mastermind class
The breaker check method returns the player is the code breaker,
while the maker check method returns the player is the code maker.

The input rounds method asks the human code maker player for the
number of rounds.

The game loop method starts the create code methods from the maker
class and the break code methods from the breaker class.